refactor: split Reflection checker metadata
