+++ src/types/checker/builtin_types/reflection/owner_helpers.rs
@@ -1,0 +1,81 @@
+//! Purpose:
+//! Builds Reflection owner constructors, attribute accessors, and variadic defaults.
+//!
+//! Called from:
+//! - The Reflection checker metadata facade and sibling builders.
+//!
+//! Key details:
+//! - Synthetic callable metadata stays aligned with direct special lowering.
+
+use super::*;
+
+/// Builds a public `__construct` method for a reflection owner class using the
+/// provided parameter list: each tuple is (name, type_expr, default, by_ref).
+pub(super) fn builtin_reflection_owner_constructor_method(
+    params: Vec<(&str, Option<TypeExpr>, Option<Expr>, bool)>,
+) -> ClassMethod {
+    let dummy_span = crate::span::Span::dummy();
+    ClassMethod {
+        name: "__construct".to_string(),
+        visibility: Visibility::Public,
+        is_static: false,
+        is_abstract: false,
+        is_final: false,
+        has_body: true,
+        params: params
+            .into_iter()
+            .map(|(name, ty, default, by_ref)| (name.to_string(), ty, default, by_ref))
+            .collect(),
+        param_attributes: Vec::new(),
+        variadic: None,
+        variadic_by_ref: false,
+        variadic_type: None,
+        return_type: None,
+        by_ref_return: false,
+        body: Vec::new(),
+        span: dummy_span,
+        attributes: Vec::new(),
+    }
+}
+
+/// Returns a public `getAttributes()` method that returns the private `__attrs`
+/// property as an `array` of `ReflectionAttribute` objects.
+pub(super) fn builtin_reflection_owner_get_attributes_method() -> ClassMethod {
+    let dummy_span = crate::span::Span::dummy();
+    ClassMethod {
+        name: "getAttributes".to_string(),
+        visibility: Visibility::Public,
+        is_static: false,
+        is_abstract: false,
+        is_final: false,
+        has_body: true,
+        params: Vec::new(),
+        param_attributes: Vec::new(),
+        variadic: None,
+        variadic_by_ref: false,
+        variadic_type: None,
+        return_type: Some(array_type()),
+        by_ref_return: false,
+        body: vec![Stmt::new(
+            StmtKind::Return(Some(Expr::new(
+                ExprKind::PropertyAccess {
+                    object: Box::new(Expr::new(ExprKind::This, dummy_span)),
+                    property: "__attrs".to_string(),
+                },
+                dummy_span,
+            ))),
+            dummy_span,
+        )],
+        span: dummy_span,
+        attributes: Vec::new(),
+    }
+}
+
+/// Marks a synthesized variadic method signature as callable with no variadic arguments.
+pub(super) fn make_reflection_variadic_optional(sig: &mut crate::types::FunctionSig) {
+    if sig.variadic.is_some() {
+        if let Some(default) = sig.defaults.last_mut() {
+            *default = empty_array();
+        }
+    }
+}
